[Refactor] Add showAlert func to extension
