+++ ClipPic/ClipPic/ViewController/Search/View/SearchResultCollectionViewCell.swift
@@ -22,7 +22,6 @@ class SearchResultCollectionViewCell: UICollectionViewCell {
         let searchResultImageView = UIImageView()
         searchResultImageView.translatesAutoresizingMaskIntoConstraints = false
         searchResultImageView.contentMode = .scaleAspectFill
-        //  searchResultImageView.layer.cornerRadius = 20
         searchResultImageView.clipsToBounds = true
         return searchResultImageView
     }()
@@ -51,7 +50,6 @@ class SearchResultCollectionViewCell: UICollectionViewCell {
     // MARK: - Methods
     
     func setUpViews() {
-        
         self.addSubview(searchResultItemView)
         searchResultItemView.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
         searchResultItemView.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
